Rozdělení, smazání mrtvého kódu
