+++ NNPTPZ1/Poly.cs
@@ -1,0 +1,103 @@
+﻿using System.Collections.Generic;
+
+namespace NNPTPZ1
+{
+
+    namespace Mathematics
+    {
+        public class Poly
+        {
+            /// <summary>
+            /// Coe
+            /// </summary>
+            public List<Cplx> Coe { get; set; }
+
+            /// <summary>
+            /// Constructor
+            /// </summary>
+            public Poly() => Coe = new List<Cplx>();
+
+            public void Add(Cplx coe) =>
+                Coe.Add(coe);
+
+            /// <summary>
+            /// Derives this polynomial and creates new one
+            /// </summary>
+            /// <returns>Derivated polynomial</returns>
+            public Poly Derive()
+            {
+                Poly p = new Poly();
+                for (int q = 1; q < Coe.Count; q++)
+                {
+                    p.Coe.Add(Coe[q].Multiply(new Cplx() { Re = q }));
+                }
+
+                return p;
+            }
+
+            /// <summary>
+            /// Evaluates polynomial at given point
+            /// </summary>
+            /// <param name="x">point of evaluation</param>
+            /// <returns>y</returns>
+            public Cplx Eval(double x)
+            {
+                var y = Eval(new Cplx() { Re = x, Imaginari = 0 });
+                return y;
+            }
+
+            /// <summary>
+            /// Evaluates polynomial at given point
+            /// </summary>
+            /// <param name="x">point of evaluation</param>
+            /// <returns>y</returns>
+            public Cplx Eval(Cplx x)
+            {
+                Cplx s = Cplx.Zero;
+                for (int i = 0; i < Coe.Count; i++)
+                {
+                    Cplx coef = Coe[i];
+                    Cplx bx = x;
+                    int power = i;
+
+                    if (i > 0)
+                    {
+                        for (int j = 0; j < power - 1; j++)
+                            bx = bx.Multiply(x);
+
+                        coef = coef.Multiply(bx);
+                    }
+
+                    s = s.Add(coef);
+                }
+
+                return s;
+            }
+
+            /// <summary>
+            /// ToString
+            /// </summary>
+            /// <returns>String repr of polynomial</returns>
+            public override string ToString()
+            {
+                string s = "";
+                int i = 0;
+                for (; i < Coe.Count; i++)
+                {
+                    s += Coe[i];
+                    if (i > 0)
+                    {
+                        int j = 0;
+                        for (; j < i; j++)
+                        {
+                            s += "x";
+                        }
+                    }
+                    if (i+1<Coe.Count)
+                    s += " + ";
+                }
+                return s;
+            }
+        }
+    }
+}
